+++ lib/punchblock/translator/asterisk/component/mrcp_recog_prompt.rb
@@ -34,6 +34,7 @@ module Punchblock
 
             raise OptionError, "A recognition-timeout value must be a positive integer."  if @recognition_timeout && @recognition_timeout < 1
             raise OptionError, "A speech-complete-timeout value must be a positive integer."  if @speech_complete_timeout && @speech_complete_timeout < 0
+            raise OptionError, "A speech-incomplete-timeout value must be -1 or a positive integer."  if @speech_incomplete_timeout && @speech_incomplete_timeout < 0
             raise OptionError, "An initial-timeout value must be -1 or a positive integer." if @initial_timeout < -1
             raise OptionError, "An inter-digit-timeout value must be -1 or a positive integer." if @inter_digit_timeout < -1
           end
@@ -53,6 +54,7 @@ module Punchblock
               opts[:sl] = input_node.sensitivity if input_node.sensitivity
               opts[:t]  = input_node.recognition_timeout if @recognition_timeout
               opts[:sct] = input_node.speech_complete_timeout if @speech_complete_timeout
+              opts[:sint] = input_node.speech_incomplete_timeout if @speech_incomplete_timeout
               yield opts
             end
           end
@@ -62,6 +64,7 @@ module Punchblock
             @inter_digit_timeout = input_node.inter_digit_timeout || -1
             @recognition_timeout = input_node.recognition_timeout
             @speech_complete_timeout = input_node.speech_complete_timeout
+            @speech_incomplete_timeout = input_node.speech_incomplete_timeout || -1
           end
 
           def grammars