+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * Bugfix: Handle a corner case crash where a recognition request is interrupted directly after a successful recognition has completed
   * Feature: Enable setting 'Dtmf-Term-Timeout' value for UniMRCP; necessary to exit sooner from successful Lumenvox recognition event
   * Feature: Enable setting 'Speech-Complete-Timeout' value for UniMRCP; necessary to allow additional input after an initial match has been recognized
+  * Feature: Enable setting 'Speech-Incomplete_timeout' value for UniMRCP; necessary to allow enough time for input if a complete match has not been recognized
 
 # [develop](https://github.com/adhearsion/punchblock)
   * Feature: Support RubySpeech builtin grammars on Asterisk and FreeSWITCH